Fix player miscalculation
halfPlayers was an int, stopping it from holding floating point numerals. halfPlayers was changed to a float to allow it to hold a percentage of a number.

+++ src/main/java/com/dinkley/jigglytatas/JigglytatasListener.java
@@ -57,7 +57,7 @@ public class JigglytatasListener implements Listener
         //Add 1 to total sleeping players, as arrays start counting from 0
         int sleepingPlayers = getSleepingPlayers(event) + 1;
         int totalPlayers = Bukkit.getOnlinePlayers().size();
-        int halfPlayers = totalPlayers / 2;
+        float halfPlayers = totalPlayers / 2;
 
         if (isNight && bedEnterResult == OK)
         {
